Five more signed bytes the retail loads prove: SB_Galleon camera/flight pattern, the KT_Rex floor-switch grace timers, the laser range bias and the ObjHits outgoing-hit suppressor
Retail load census per struct offset, every load in the whole DOL:

  SBGalleonState 0x70 cameraState        4 loads, 4 lbz+extsb, and two of
                                        them feed signed bge/ble compares
  SBGalleonState 0x7A flightPattern      4 loads, 4 lbz+extsb
  KtRexFloorSwitch 0x04 graceTimer       5 loads, 3 lbz+extsb, 2 truncated
                                        straight back into a stb
  KtRexFloorSwitch 0x05 prevGraceTimer   1 load, lbz+extsb
  LaserBeamState 0x26 rangeOffset        1 load, lbz+extsb feeding the
                                        signed int->float conversion
  ObjHitsPriorityState 0x70              6 loads, 5 lbz+extsb, 1 stored back
    suppressOutgoingHits

None of the six is read zero-extended anywhere: no cmplwi, no andi, no
clrlwi, no unsigned widening. Retype them s8 and drop the seventeen reader
casts and address puns that were carrying the signedness by hand.

Every object byte-identical; main.dol OK.

+++ src/dlls/objects/260_SmallBasket/SmallBasket.c
@@ -485,7 +485,7 @@ int SmallBasket_resolveCollision(GameObject* obj) {
         return 1;
     }
 
-    if ((int)(hitState->objectHitMask >> 4) != 0 && (s8)hitState->suppressOutgoingHits == 0) {
+    if ((int)(hitState->objectHitMask >> 4) != 0 && hitState->suppressOutgoingHits == 0) {
         endPoints[0] = obj->anim.localPosX;
         *(endPointY = &endPoints[1]) = obj->anim.localPosY;
         *(endPointZ = &endPoints[2]) = obj->anim.localPosZ;
